2_DL_Detection/requirements.txt
absl-py==1.3.0
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==5.3.0
altgraph==0.17.3
amqp==5.2.0
annotated-types==0.6.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asttokens==2.1.0
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.3
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bidict==0.23.1
billiard==4.2.0
bleach==5.0.1
blinker==1.7.0
boltiot==1.11.2
boto==2.49.0
bs4==0.0.1
CacheControl==0.14.0
cachetools==5.2.0
celery==5.4.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.1
comtypes==1.1.14
contourpy==1.0.6
cryptography==42.0.8
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
distro==1.8.0
Django==4.2.7
django-cors-headers==4.4.0
django-csv==1.2
django-pdf==1.0.2
djangorestframework==3.15.2
dnspython==2.7.0
docker==7.1.0
dpkt==1.9.8
easyocr==1.7.2
entrypoints==0.4
ERDot==2.3.1
et-xmlfile==1.1.0
eventlet==0.38.2
exceptiongroup==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.8.0
firebase-admin==6.5.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.5.1
flatbuffers==24.3.25
fonttools==4.38.0
fpdf==1.7.2
frozendict==2.4.4
frozenlist==1.4.0
fsspec==2023.12.2
gast==0.4.0
gcloud==0.18.3
ghettoq==0.4.5
gitdb==4.0.11
GitPython==3.1.43
google==3.0.0
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
greenlet==3.1.1
grpcio==1.65.0
grpcio-status==1.62.2
gTTS==2.5.4
h11==0.14.0
h5py==3.11.0
html5lib==1.1
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.2
huggingface-hub==0.20.1
idna==3.4
ifaddr==0.2.0
ImageHash==4.3.1
imageio==2.22.4
imutils==0.5.4
iniconfig==2.0.0
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
itsdangerous==2.1.2
jax==0.4.30
jaxlib==0.4.30
jedi==0.18.1
jellyfish==1.0.0
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.10
jsonschema==4.17.0
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-server==1.23.1
jupyter_client==7.4.4
jupyter_core==5.7.1
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.16.2
jwcrypto==1.5.6
keras==3.4.1
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
kombu==5.3.7
libclang==14.0.6
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.14
mistune==2.0.4
ml-dtypes==0.3.2
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.8
mtcnn==0.1.1
multidict==6.0.4
multitasking==0.0.11
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==9.0.0
mysqlclient==2.1.1
namex==0.0.8
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.4
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.8
ngrok==1.4.0
ninja==1.11.1.3
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.3.6
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.12.0
outcome==1.2.0
packaging==21.3
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.3
paypalrestsdk==1.13.3
peewee==3.17.6
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.2.0
pipenv==2022.11.30
platformdirs==2.5.3
playsound==1.3.0
plotly==5.22.0
pluggy==1.3.0
prometheus-client==0.15.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
psutil==5.9.4
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.14
PyAutoGUI==0.9.54
pyclipper==1.3.0.post6
pycoingecko==3.1.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.9.1
pygame==2.6.1
pygeoip==0.3.2
PyGetWindow==0.0.9
Pygments==2.13.0
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.3
PyJWT==2.8.0
PyMsgBox==1.0.9
pyngrok==7.2.0
pynput==1.7.7
pyOpenSSL==24.2.1
pyparsing==3.0.9
PyPDF2==3.0.1
pyperclip==1.9.0
pypiwin32==223
Pyrebase4==4.8.0
PyRect==0.2.0
pyrsistent==0.19.2
PyScreeze==0.1.30
PySocks==1.7.1
pytesseract==0.3.13
pytest==7.4.3
python-bidi==0.6.3
python-dateutil==2.8.2
python-decouple==3.8
python-engineio==4.11.2
python-http-client==3.3.7
python-jwt==4.1.0
python-socketio==5.12.1
pyttsx3==2.90
pytweening==1.2.0
pytz==2023.3
PyWavelets==1.4.1
pywin32==304
pywin32-ctypes==0.2.2
pywinpty==2.0.9
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.5.1
QtPy==2.4.1
regex==2023.10.3
reportlab==4.0.4
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rich==13.7.1
rsa==4.9
safetensors==0.4.1
scapy==2.5.0
scikit-image==0.19.3
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
selenium==4.8.0
Send2Trash==1.8.0
sendgrid==6.10.0
shapely==2.0.6
sib-api-v3-sdk @ git+https://github.com/sendinblue/APIv3-python-library.git@9a5807c149db7855eca346b1d7829a96481f495c
simple-websocket==1.1.0
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.7
soupsieve==2.3.2.post1
SpeechRecognition==3.10.1
sqlparse==0.4.3
stack-data==0.6.0
starkbank-ecdsa==2.2.0
streamlit==1.36.0
sympy==1.13.1
tenacity==8.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.16.2
tensorflow-estimator==2.10.0
tensorflow-intel==2.16.2
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
terminado==0.17.0
threadpoolctl==3.3.0
tifffile==2022.10.10
tinycss2==1.2.1
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchvision==0.20.1+cu124
tornado==6.2
tqdm==4.67.1
traitlets==5.5.0
transformers==4.36.2
trio==0.22.0
trio-websocket==0.9.2
twilio==8.7.0
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.3.58
ultralytics-thop==2.0.13
uritemplate==4.1.1
urllib3==1.26.12
vine==5.1.0
virtualenv==20.17.0
virtualenv-clone==0.5.7
watchdog==4.0.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.2
yfinance==0.2.40
zeroconf==0.136.0
